requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.1.2
coverage==6.3.2
flake8==4.0.1
Flask==2.1.1
Flask-Cors==3.0.10
flask-restx==0.5.1
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.1
pytest-cov==3.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-mock==3.7.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
smmap==5.0.0
tomli==2.0.1
Werkzeug==2.1.1
zipp==3.8.0
